+++ csalt/c_uint32.c
@@ -19,9 +19,9 @@
 // ================================================================================ 
 // ================================================================================ 
 
-uint32_tensor_expect_t init_uint32_tensor(uint32_t            ndim,
-                                        const size_t*      shape,
-                                        allocator_vtable_t alloc_v) {
+uint32_tensor_expect_t init_uint32_tensor(uint8_t            ndim,
+                                          const size_t*      shape,
+                                          allocator_vtable_t alloc_v) {
     /* Allocate the underlying tensor */
     tensor_expect_t r = init_tensor(ndim, shape, UINT32_TYPE, alloc_v);
     if (!r.has_value)
